+++ src/pictures.js
@@ -1,14 +1,7 @@
 'use strict';
 
-(function() {
-
-// define(function() {
-//  require('script');
-// });
-define('filterPic', ['filter'], function() {
-});
-require(['filterPic'], function(filterPic) {});
-
+define(['filter', 'ajax'], function(getFilteredPictures, getPictures) {
+  console.log(getFilteredPictures);
   var picturesContainer = document.querySelector('.pictures');
   var containerSides = picturesContainer.getBoundingClientRect();
   var templateElement = document.querySelector('#picture-template');
@@ -16,7 +9,6 @@ require(['filterPic'], function(filterPic) {});
   var pics = [];
   var filteredPictures = [];
   var elementToClone;
-  var LOAD_URL = '//o0.github.io/assets/json/pictures.json';
   var PAGE_SIZE = 12;
   var pageNumber = 0;
 
@@ -95,31 +87,6 @@ require(['filterPic'], function(filterPic) {});
     renderNextPages();
   };
 
-  var getFilteredPictures = function(pictures, filter) {
-    var picturesToFilter = pictures.slice(0);
-
-    switch(filter) {
-      case 'filter-popular':
-        break;
-      case 'filter-new':
-        picturesToFilter = picturesToFilter.filter(function(elem) {
-          var dateTwoWeeksAgo = new Date(elem.date);
-          var nowDate = new Date();
-          return dateTwoWeeksAgo > nowDate - 14 * 24 * 60 * 60 * 1000;
-        });
-        picturesToFilter = picturesToFilter.sort(function(a, b) {
-          return new Date(b.date) - new Date(a.date);
-        });
-        break;
-      case 'filter-discussed':
-        picturesToFilter = picturesToFilter.sort(function(a, b) {
-          return b.comments - a.comments;
-        });
-        break;
-    }
-    return picturesToFilter;
-  };
-
   var setFilterEnabled = function(filter) {
     filteredPictures = getFilteredPictures(pics, filter);
     pageNumber = 0;
@@ -132,23 +99,6 @@ require(['filterPic'], function(filterPic) {});
         setFilterEnabled(evt.target.id);
       }
     });
-  };
-
-  var getPictures = function(callback) {
-    var xhr = new XMLHttpRequest();
-
-    xhr.onload = function(evt) {
-      var loadedData = JSON.parse(evt.target.response);
-      picturesContainer.classList.add('pictures-loading');
-      callback(loadedData);
-    };
-
-    xhr.onerror = function() {
-      picturesContainer.classList.add('pictures-failure');
-    };
-
-    xhr.open('GET', LOAD_URL);
-    xhr.send();
   };
 
   var setScrollEnabled = function() {
@@ -175,4 +125,5 @@ require(['filterPic'], function(filterPic) {});
   });
 
   filters.classList.remove('hidden');
-})();
+});
+
